refactor: emergencyCenter type 리팩토링 및 컴포넌트 타입 리팩토링, console log제거
