+++ src/components/prototypes/EmergencyCenter/Detail/index.tsx
@@ -1,12 +1,26 @@
 import { EmergencyCenterDetailContainer } from "@/components/container/EmergencyCenterCentainer/Detail";
-import { GetEmergencyCenterDetailResponse } from "@/types/emergencyCenter.type";
+import { Department } from "@/types/department/client";
+import {
+  EmergencyCenter,
+  EmergencyRoomWithBeds,
+} from "@/types/model/emergencyCenter";
 
 interface EmergencyCenterDetailProps {
-  emergency_center: GetEmergencyCenterDetailResponse; // ssr 응답값을 직접 사용
+  emergencyCenter: EmergencyCenter;
+  emergencyRooms: EmergencyRoomWithBeds[];
+  departments: Department[];
 }
 
 export const EmergencyCenterDetailPrototype = ({
-  emergency_center,
+  emergencyCenter,
+  emergencyRooms,
+  departments,
 }: EmergencyCenterDetailProps) => {
-  return <EmergencyCenterDetailContainer emergency_center={emergency_center} />;
+  return (
+    <EmergencyCenterDetailContainer
+      emergencyCenter={emergencyCenter}
+      emergencyRooms={emergencyRooms}
+      departments={departments}
+    />
+  );
 };
