+++ src/hooks/api/api.ts
@@ -71,7 +71,6 @@ export function useGetApi<P extends MethodPaths<"get">>(
     return data;
   });
   const _data = getData as SuccessResponse<P, "get">;
-  // console.log("getApi", url, loader.id, _data, _error, isLoadingSWR);
 
   useEffect(() => {
     if (!isLoadingSWR && (_data || _error))
