+++ src/hooks/useLoading.tsx
@@ -19,11 +19,9 @@ export function useLoading() {
   return useMemo(
     () => ({
       on: () => {
-        // console.log("on:", id);
         context.on(id);
       },
       off: () => {
-        // console.log("off:", id);
         context.off(id);
       },
       // id,
