+++ src/types/department/client.ts
@@ -2,5 +2,5 @@ export interface Department {
   department_id: number;
   department_name: string;
   parent_department_id?: number | null;
-  sub_departments: Department[];
+  status: "ACTIVE" | "INACTIVE" | "DELETED";
 }
